Cleaned up the launch activity and added a splash screen
