+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name=".SplashScreen"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -31,7 +31,7 @@
             android:label="Sign up"
             android:parentActivityName=".LoginActivity" />
         <activity android:name=".SettingsActivity" />
-        <activity android:name=".FragmentedActivity" />
+        <activity android:name=".MainActivity" />
     </application>
 
 </manifest>
