+++ app/src/main/java/acc/aviato/LoginActivity.java
@@ -1,11 +1,8 @@
 package acc.aviato;
 
-import android.app.Activity;
 import android.app.AlertDialog;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
@@ -14,9 +11,7 @@ import android.widget.Button;
 import android.widget.EditText;
 
 import com.parse.LogInCallback;
-import com.parse.Parse;
 import com.parse.ParseException;
-import com.parse.ParseObject;
 import com.parse.ParseUser;
 
 public class LoginActivity extends AppCompatActivity {
@@ -36,7 +31,7 @@ public class LoginActivity extends AppCompatActivity {
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
-        mSignUp = (Button)findViewById(R.id.signUpButton);
+        mSignUp = (Button) findViewById(R.id.signUpButton);
         mSignUp.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -45,17 +40,17 @@ public class LoginActivity extends AppCompatActivity {
             }
         });
 
-        mPhoneNumber = (EditText)findViewById(R.id.phoneNumberField);
-        mPassword = (EditText)findViewById(R.id.passwordField);
-        mSkip = (Button)findViewById(R.id.skipButton);
+        mPhoneNumber = (EditText) findViewById(R.id.phoneNumberField);
+        mPassword = (EditText) findViewById(R.id.passwordField);
+        mSkip = (Button) findViewById(R.id.skipButton);
         mSkip.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(LoginActivity.this, FragmentedActivity.class);
+                Intent intent = new Intent(LoginActivity.this, MainActivity.class);
                 startActivity(intent);
             }
         });
-        mLoginButton = (Button)findViewById(R.id.loginButton);
+        mLoginButton = (Button) findViewById(R.id.loginButton);
         mLoginButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -72,8 +67,7 @@ public class LoginActivity extends AppCompatActivity {
                             .setPositiveButton(android.R.string.ok, null);
                     AlertDialog dialog = builder.create();
                     dialog.show();
-                }
-                else {
+                } else {
                     // Login
                     setProgressBarIndeterminateVisibility(true);
 
@@ -84,7 +78,7 @@ public class LoginActivity extends AppCompatActivity {
 
                             if (e == null) {
                                 // Success!
-                                Intent intent = new Intent(LoginActivity.this, FragmentedActivity.class);
+                                Intent intent = new Intent(LoginActivity.this, MainActivity.class);
                                 intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
                                 intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK);
                                 startActivity(intent);
