+++ app/src/main/java/acc/aviato/SplashScreen.java
@@ -1,0 +1,41 @@
+package acc.aviato;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.os.Handler;
+import android.support.v4.view.GravityCompat;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+import android.view.MenuItem;
+
+import com.parse.ParseUser;
+
+public class SplashScreen extends AppCompatActivity {
+
+    private final int SPLASH_DISPLAY_LENGTH = 1000;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_splashscreen);
+
+        new Handler().postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                // Checks if the user is not signed in
+                if (ParseUser.getCurrentUser() == null) {
+                    Intent intent = new Intent(SplashScreen.this, LoginActivity.class);
+                    intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                    startActivity(intent);
+                    SplashScreen.this.finish();
+                } else {
+                    Intent intent = new Intent(SplashScreen.this, MainActivity.class);
+                    intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                    startActivity(intent);
+                    SplashScreen.this.finish();
+                }
+            }
+        }, SPLASH_DISPLAY_LENGTH);
+    }
+}
