+++ app/src/main/res/layout/activity_sign_up.xml
@@ -1,42 +1,47 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
-    android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    android:paddingBottom="@dimen/activity_vertical_margin"
     tools:context="acc.aviato.SignUpActivity">
 
     <EditText
+        android:id="@+id/phoneNumberField"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:inputType="phone"
-        android:ems="10"
-        android:id="@+id/phoneNumberField"
         android:layout_alignParentTop="true"
         android:layout_centerHorizontal="true"
-        android:layout_marginTop="122dp" />
+        android:layout_marginTop="122dp"
+        android:ems="10"
+        android:hint="Enter cell # 9797905555"
+        android:inputType="phone" />
 
     <EditText
+        android:id="@+id/passwordField"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:inputType="textPassword"
-        android:ems="10"
-        android:id="@+id/passwordField"
         android:layout_below="@+id/phoneNumberField"
-        android:layout_centerHorizontal="true" />
+        android:layout_centerHorizontal="true"
+        android:ems="10"
+        android:hint="Enter password"
+        android:inputType="textPassword" />
 
     <Button
+        android:id="@+id/signUpButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="New Button"
-        android:id="@+id/signUpButton"
-        android:layout_below="@+id/passwordField"
-        android:layout_marginTop="40dp"
-        android:layout_alignRight="@+id/passwordField"
         android:layout_alignEnd="@+id/passwordField"
         android:layout_alignLeft="@+id/passwordField"
+        android:layout_alignRight="@+id/passwordField"
         android:layout_alignStart="@+id/passwordField"
+        android:layout_below="@+id/passwordField"
+        android:layout_marginTop="40dp"
         android:background="@android:color/holo_blue_light"
-        android:textColor="#ffffffff" />
+        android:text="Sign Up"
+        android:textColor="#ffffffff"
+        android:textSize="20sp" />
 </RelativeLayout>
