+++ app/src/main/res/layout/content_login.xml
@@ -17,7 +17,8 @@
         android:id="@+id/phoneNumberField"
         android:layout_alignParentTop="true"
         android:layout_centerHorizontal="true"
-        android:layout_marginTop="123dp" />
+        android:layout_marginTop="123dp"
+        android:hint="Enter cell # 9797905555" />
 
     <EditText
         android:layout_width="wrap_content"
@@ -26,7 +27,8 @@
         android:ems="10"
         android:id="@+id/passwordField"
         android:layout_below="@+id/phoneNumberField"
-        android:layout_centerHorizontal="true" />
+        android:layout_centerHorizontal="true"
+        android:hint="Enter password"/>
 
     <Button
         android:layout_width="wrap_content"
